+++ virbius-kernel/plugin/json_unmarshal.go
@@ -1,7 +1,0 @@
-package main
-
-import "encoding/json"
-
-func jsonUnmarshalImpl(data []byte, v any) error {
-	return json.Unmarshal(data, v)
-}